Moved somePromises to util
